Moved projector code to a mixin, issue #169
