+++ tests/unit/mixins/all.ts
@@ -1,5 +1,6 @@
 import './createCssTransitionMixin';
 import './createFormFieldMixin';
 import './createI18nMixin';
+import './createProjectorMixin';
 import './shallowPropertyComparisonMixin';
 import './externalState';
